requirements.txt
django==5.0.6
django-phonenumber-field[phonenumberslite]
crispy-bootstrap5==2024.2